requirements_list.txt
Django==5.2
channels==4.2.2
channels-redis==4.2.1
daphne==4.1.2
djangorestframework==3.16.0
redis==5.0.1
Pillow==10.2.0
aioredis==2.0.1
asgiref==3.7.2
async-timeout==4.0.3
autobahn==23.6.2
Automat==22.10.0
cryptography==41.0.7
hiredis==2.2.3
hyperlink==21.0.0
incremental==22.10.0
msgpack==1.0.7
pytz==2024.1
sqlparse==0.4.4
txaio==23.1.1
zope.interface==6.1